refactor(db): split shared domain types
