+++ app/lib/db/project.ts
@@ -1,0 +1,15 @@
+export interface Project {
+  _id: string;
+  _creationTime: number;
+  name: string;
+  travelDestination?: string;
+  travelPurpose?: string;
+  organizationId: string;
+  isArchived: boolean;
+  createdBy: string;
+}
+
+export type ProjectTravelDefaults = Pick<
+  Project,
+  "travelDestination" | "travelPurpose"
+>;
