+++ app/lib/db/reimbursement.ts
@@ -1,0 +1,72 @@
+export type ReimbursementType = "expense" | "travel";
+export type ReviewStatus =
+  | "pending"
+  | "changes_requested"
+  | "approved"
+  | "declined";
+export type CostType =
+  | "car"
+  | "train"
+  | "flight"
+  | "taxi"
+  | "bus"
+  | "accommodation";
+
+export interface Reimbursement {
+  _id: string;
+  _creationTime: number;
+  organizationId: string;
+  projectId: string;
+  amount: number;
+  type: ReimbursementType;
+  status: ReviewStatus;
+  iban: string;
+  bic?: string;
+  accountHolder: string;
+  rejectionNote?: string;
+  reviewNote?: string;
+  createdBy: string;
+  reviewedBy?: string;
+  reviewedAt?: number;
+  currency?: string;
+  signatureStorageId?: string;
+  isSharedLink?: boolean;
+  submitterName?: string;
+  submitterEmail?: string;
+  invitedName?: string;
+  invitedEmail?: string;
+  submittedExternally?: boolean;
+  requestedExternally?: boolean;
+  description?: string;
+  pendingUploadKeys?: string[];
+}
+
+export interface TravelDetails {
+  _id: string;
+  _creationTime: number;
+  reimbursementId: string;
+  startDate: string;
+  endDate: string;
+  destination: string;
+  purpose: string;
+  isInternational: boolean;
+  mealAllowanceDays?: number;
+  mealAllowanceDailyBudget?: number;
+  allowFoodAllowance?: boolean;
+}
+
+export interface Receipt {
+  _id: string;
+  _creationTime: number;
+  reimbursementId: string;
+  receiptNumber?: string;
+  receiptDate: string;
+  companyName: string;
+  description: string;
+  netAmount: number;
+  taxRate: number;
+  grossAmount: number;
+  fileStorageId: string;
+  costType?: CostType;
+  kilometers?: number;
+}
